social-drawing/requirements.txt
appdirs==1.4.3
arabic-reshaper==3.0.0
astunparse==1.6.3
async-generator==1.10
attrdict==2.0.1
attrs==23.1.0
bcrypt==3.1.7
beautifulsoup4==4.12.2
blinker==1.4
blosc2==2.0.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
Click==7.0
colorama==0.4.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
Cython==0.29.34
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.4
demjson==2.2.4
demjson3==3.0.6
distlib==0.3.0
distro==1.8.0
dukpy==0.3.0
egi-pynetstation==0.0.0rc1
elementpath==4.1.5
entrypoints==0.4
esprima==4.0.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fasteners==0.14.1
ffpyplayer==4.5.0
filelock==3.0.12
fonttools==4.39.4
freetype-py==2.4.0
future==0.18.3
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
html2text==2020.1.16
httplib2==0.14.0
icub-models==2.3.0
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==1.5.0
importlib-resources==5.12.0
javascripthon==0.12
jedi==0.18.2
jeepney==0.8.0
json-tricks==3.16.1
jtutils==0.0.8
keyring==18.0.1
kiwisolver==1.4.4
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
leven==1.0.4
lockfile==0.12.2
louis==1.3
macaroonbakery==1.3.1
Mako==1.1.0
markdown-it-py==2.2.0
MarkupSafe==1.1.0
matplotlib==3.7.1
mdurl==0.1.2
monotonic==1.5
more-itertools==4.2.0
moviepy==1.0.3
msgpack==1.0.5
msgpack-numpy==0.4.8
ndjson==0.3.1
netifaces==0.10.4
nose==1.3.7
ntplib==0.4.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
panda==0.3.1
pandas==2.0.1
parso==0.8.3
pbr==5.11.1
Pillow==9.5.0
proglog==0.1.10
protobuf==3.20.3
psutil==5.9.5
PsychoPy==2023.1.2
psychtoolbox==3.0.19.0
py-cpuinfo==9.0.0
pycparser==2.21
pyglet==1.4.11	
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparallel==0.2.2
pyparsing==3.0.9
pypi-search==1.2.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
python-bidi==0.4.2
python-csv==0.0.13
python-dateutil==2.8.2
python-gitlab==3.14.0
python-statemachine==2.0.0
python-vlc==3.0.11115
python-xlib==0.33
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
questplus==2023.1
requests==2.30.0
requests-toolbelt==1.0.0
scipy==1.10.1
screeninfo==0.8.1
SecretStorage==3.3.3
selenium==4.9.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
tables==3.8.0
testresources==2.0.1
tk==0.1.0
tobii-research==1.10.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.2
wadllib==1.3.6
websocket-client==1.5.1
websockets==12.0
wsproto==1.2.0
wxPython==4.2.0
xarray==2023.1.0
xlrd==2.0.1
xmlschema==2.5.0
xmltodict==0.13.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0


